+++ dcplib/checksumming_io/checksumming_sink.py
@@ -9,7 +9,7 @@ A file-like object that computes checksums for the data written to it, discardin
 
 class ChecksummingSink:
 
-    def __init__(self, hash_functions=('crc32c', 'sha1', 'sha256', 's3_etag')):
+    def __init__(self, write_chunk_size, hash_functions=('crc32c', 'sha1', 'sha256', 's3_etag')):
         self._hashers = dict()
         for hasher in hash_functions:
             if hasher == 'crc32c':
@@ -19,7 +19,7 @@ class ChecksummingSink:
             elif hasher == 'sha256':
                 self._hashers['sha256'] = hashlib.sha256()
             elif hasher == 's3_etag':
-                self._hashers['s3_etag'] = S3Etag()
+                self._hashers['s3_etag'] = S3Etag(write_chunk_size)
 
     def write(self, data):
         for hasher in self._hashers.values():